backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiofiles
python-dotenv
httpx
yfinance
pandas
pandas-ta
pydantic
slowapi
